fixes #35
Moved portfolio methods to protected and implemented portfolio service
that is injected into the strategy as the interface to access the
postions, value and balances of a strategy. Fill was updated to not set
the time as this was causing issues with replay and rolling the
epruntime to the current date.